Extract TransactionLogFilesSnapshot
